+++ src/main/java/roomescape/controller/AdminReservationTimeController.java
@@ -1,0 +1,37 @@
+package roomescape.controller;
+
+import java.net.URI;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.dto.ReservationTimeRequestDTO;
+import roomescape.dto.ReservationTimeResponseDTO;
+import roomescape.service.ReservationTimeService;
+
+@RestController
+@RequestMapping("/admin")
+public class AdminReservationTimeController {
+
+    private final ReservationTimeService reservationTimeService;
+
+    public AdminReservationTimeController(ReservationTimeService reservationTimeService) {
+        this.reservationTimeService = reservationTimeService;
+    }
+
+    @PostMapping("/times")
+    public ResponseEntity<Void> add(
+            @RequestBody ReservationTimeRequestDTO request) {
+        ReservationTimeResponseDTO saved = reservationTimeService.addReservationTime(request);
+        return ResponseEntity.created(URI.create("/times/" + saved.id())).build();
+    }
+
+    @DeleteMapping("/times/{id}")
+    public ResponseEntity<Void> deleteReservationTime(@PathVariable Long id) {
+        reservationTimeService.deleteReservationTime(id);
+        return ResponseEntity.noContent().build();
+    }
+}
