+++ src/main/java/roomescape/controller/AdminThemeController.java
@@ -1,0 +1,37 @@
+package roomescape.controller;
+
+import java.net.URI;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.dto.ThemeRequestDTO;
+import roomescape.dto.ThemeResponseDTO;
+import roomescape.service.ThemeService;
+
+@RestController
+@RequestMapping("/admin")
+public class AdminThemeController {
+
+    private final ThemeService themeService;
+
+    public AdminThemeController(ThemeService themeService) {
+        this.themeService = themeService;
+    }
+
+    @PostMapping("/themes")
+    public ResponseEntity<Void> add(
+            @RequestBody ThemeRequestDTO request) {
+        ThemeResponseDTO saved = themeService.addTheme(request);
+        return ResponseEntity.created(URI.create("/themes/" + saved.id())).build();
+    }
+
+    @DeleteMapping("/themes/{id}")
+    public ResponseEntity<Void> deleteById(@PathVariable Long id) {
+        themeService.deleteTheme(id);
+        return ResponseEntity.noContent().build();
+    }
+}
